Store memory initialization in string literal if requested
With “--memory-init-file 0 -s MEM_INIT_METHOD=2” on the command line, the
generated JavaScript file will contain a string literal representing the
initial content of the memory buffer.

The MEM_INIT_METHOD defaults to 0 but gets set to 1 if --memory-init-file is
being used.  Conversely, setting it to 1 without --memory-init-file will
cause it to revert to 0.  That way, we can use the setting in the postamble,
without too many changes in other places.

Since memory is initialized to all zero, trailing zeros can be omitted.
This change affects the file-based initialization as well, but not when
using emterpretify.