Adapt test runner to memory initializer strings
The assumption that “-O2” implies “--memory-init-file 1” only holds unless
overridden.  Since we now have a whole test run where this is not the case,
we need to take such an override into account.  Refactored existing code
into a resuable method.

The assumption that “'/*' not in src” implies “cleaned-up sources with no
comments” is no longer true, since we may well have a literal “/*” inside
the memory initializer string.  In fact we have just this for several
existing test cases, e.g. fnmatch or strings.  I dropped that assertion
altogether, since properly fixing it felt more complex (and therefore less
reliable) than the thing it's supposed to test for.